Created a script based off of the original build_waggle_image.py that fetches the waggle base image (not the odroid stock image), pulls appropriate waggle repositories for the particular odoid type, and then configures each waggle repository.

+++ build_image.py
@@ -6,45 +6,52 @@ import uuid
 
 
 # To copy a new public image to the download webpage, copy the waggle-id_rsa ssh key to /root/. 
-# To generate a functional AoT image with private configuration, put id_rsa_waggle_aot_config and a clone of git@gith_Aub.com:waggle-sensor/private_config.git in /root
+# To generate a functional AoT image with private configuration, put id_rsa_waggle_aot_config and a clone of git@github.com:waggle-sensor/private_config.git in /root
 
 # One of the most significant modifications that this script does is setting static IPs. Nodecontroller and guest node have different static IPs.
 
 
-print "usage: python -u ./build_waggle_image.py 2>&1 | tee build.log"
+print "usage: python -u ./build_image.py 2>&1 | tee build.log"
 
 start_time = time.time()
 
 debug=0 # skip chroot environment if 1
 
+build_uuid = uuid.uuid1()
+
 waggle_image_directory = os.path.dirname(os.path.abspath(__file__))
-print("### Run directory for build_waggle_image.py: %s" % waggle_image_directory)
+print("### Run directory for build_image.py: %s" % waggle_image_directory)
 data_directory="/root"
+
+uuid_file = '%s/build_uuid' % data_directory
+with open(uuid_file, 'w') as uuid:
+  uuid.write(str(build_uuid))
 
 report_file="/root/report.txt"
 
-waggle_stock_url='http://www.mcs.anl.gov/research/projects/waggle/downloads/stock_images/'
+waggle_stock_url='http://www.mcs.anl.gov/research/projects/waggle/downloads/waggle_images/base/'
 base_images=   {
-                'odroid-xu3' : {
-                        'filename': "ubuntu-14.04lts-server-odroid-xu3-20150725.img",
+                'odroid-xu4' : {
+                        'filename': "waggle-base-extension_node-odroid-xu4-20160926.img",
                          'url': waggle_stock_url
                         },
                 'odroid-c1' : {
-                        'filename':"ubuntu-14.04.3lts-lubuntu-odroid-c1-20151020.img",
+                        'filename':"waggle-base-nodecontroller-odroid-c1-20160926.img",
                         'url': waggle_stock_url
                     } 
                 }
 
 
-create_b_image = 0 # will be 1 for XU3/4
+create_b_image = 0 # will be 1 for XU4
 
 change_partition_uuid_script = waggle_image_directory + '/change_partition_uuid.sh'   #'/usr/lib/waggle/waggle_image/change_partition_uuid.sh'
 
-mount_point="/mnt/newimage"
+mount_point_A="/mnt/newimage_A"
+mount_point_B="/mnt/newimage_B"
 
 
 
-is_extension_node = 0 # will be set automatically to 1 if an odroid-xu3 is detected !
+is_extension_node = 0 # will be set automatically to 1 if an odroid-xu4 is detected !
 
 
 
@@ -129,22 +136,6 @@ def write_file(filename, content):
     print "writing file "+filename
     with open(filename, "w") as text_file:
         text_file.write(content)
-
-
-def write_build_script(filename, node_script_filename):
-  base_build_init_script_filename = waggle_image_directory + '/scripts/base_build_init.in'
-  with open(base_build_init_script_filename) as script:
-    init_script = script.read()
-
-  abs_node_script_filename = waggle_image_directory + '/scripts/' + node_script_filename
-  with open(abs_node_script_filename) as script:
-    node_script = script.read()
-
-  base_build_final_script_filename = waggle_image_directory + '/scripts/base_build_final.in'
-  with open(base_build_final_script_filename) as script:
-    final_script = script.read()
-
-  write_file(filename, init_script+node_script+final_script)
             
 
 def unmount_mountpoint(mp):
@@ -197,7 +188,7 @@ def create_loop_devices(filename, device_number, start_block_boot, start_block_d
     loop_partition_2 = loop_device+'p2' # data/root partition
     
     if not start_block_data:
-        # example: fdisk -lu waggle-extension_node-odroid-xu3-20160601.img | grep "^waggle-extension_node-odroid-xu3-20160601.img2" | awk '{{print $2}}'
+        # example: fdisk -lu waggle-extension_node-odroid-xu4-20160601.img | grep "^waggle-extension_node-odroid-xu4-20160601.img2" | awk '{{print $2}}'
         start_block_data_str = get_output("fdisk -lu {0} | grep '{0}2' | awk '{{print $2}}'".format(filename))
         start_block_data=int(start_block_data_str)
         print "start_block_data: ", start_block_data
@@ -242,11 +233,11 @@ def detect_odroid_model():
     odroid_model_raw=get_output("head -n 1 /media/boot/boot.ini | cut -d '-' -f 1 | tr -d '\n'")
     odroid_model=""
     
-    # The XU4 is actually a XU3.
+    # The XU4 is actually an XU3.
     if odroid_model_raw == "ODROIDXU":
         print "Detected device: %s" % (odroid_model_raw)
         if os.path.isfile('/media/boot/exynos5422-odroidxu3.dtb'):
-            odroid_model="odroid-xu3"
+            odroid_model="odroid-xu4"
             #is_extension_node = 1
         else:
             odroid_model="odroid-xu"
@@ -278,7 +269,7 @@ def min_used_minor(device_minor_used):
 
 # clean up first
 
-unmount_mountpoint(mount_point)
+unmount_mountpoint(mount_point_A)
 
     
 time.sleep(3)
@@ -327,7 +318,7 @@ if not odroid_model:
     sys.exit(1)
 
 
-if odroid_model == "odroid-xu3":
+if odroid_model == "odroid-xu4":
     is_extension_node = 1
     create_b_image = 1
 
@@ -343,10 +334,12 @@ else:
 
 print "image_type: ", image_type
 
-new_image_base="waggle-base-%s-%s-%s" % (image_type, odroid_model, date_today) 
+new_image_base="waggle-%s-%s-%s" % (image_type, odroid_model, date_today) 
 new_image_prefix="%s/%s" % (data_directory, new_image_base)
-new_image="%s.img" % (new_image_prefix)
-new_image_compressed = new_image+'.xz'
+new_image_a="%s.img" % (new_image_prefix)
+new_image_a_compressed = new_image_a+'.xz'
+
+new_image_b="%s_B.img" % (new_image_prefix)
 
 os.chdir(data_directory)
 
@@ -381,12 +374,12 @@ print("Base Image Unpacking Duration: %ds" % (image_unpack_time - image_fetch_ti
 ####################
 
 try:
-    os.remove(new_image)
+    os.remove(new_image_a)
 except:
     pass
 
-print "Copying file %s to %s ..." % (base_image, new_image)
-shutil.copyfile(base_image, new_image)
+print "Copying file %s to %s ..." % (base_image, new_image_a)
+shutil.copyfile(base_image, new_image_a)
 
 ###### TIMING ######
 image_copy_time = time.time()
@@ -397,7 +390,7 @@ print("Base Image Copy Duration: %ds" % (image_copy_time - image_unpack_time))
 # LOOP DEVICES HERE
 #
 
-start_block_boot, start_block_data = create_loop_devices(new_image, 0, None, None)
+start_block_boot, start_block_data = create_loop_devices(new_image_a, 0, None, None)
 
 
 
@@ -407,52 +400,47 @@ check_partition(0)
 
 
 
-print "execute: mkdir -p "+mount_point
+print "execute: mkdir -p "+mount_point_A
 try: 
-    os.mkdir(mount_point)
+    os.mkdir(mount_point_A)
 except:
     pass
 
 
-mount_mountpoint(0, mount_point)
+print "execute: mkdir -p "+mount_point_B
+try: 
+    os.mkdir(mount_point_B)
+except:
+    pass
+
+
+mount_mountpoint(0, mount_point_A)
 
 
 # TODO remove this test
-unmount_mountpoint(mount_point)
+unmount_mountpoint(mount_point_A)
 time.sleep(3)
 destroy_loop_devices()
 time.sleep(2)
-create_loop_devices(new_image, 0,  None, start_block_data)
+create_loop_devices(new_image_a, 0,  None, start_block_data)
 print "filesystem check on /dev/loop0p2 after first mount"
 check_partition(0)
 
 
 
-mount_mountpoint(0, mount_point)
+mount_mountpoint(0, mount_point_A)
 
 ###### TIMING ######
 loop_mount_time = time.time()
 print("Loop Mount Duration: %ds" % (loop_mount_time - image_copy_time))
 ####################
 
-shutil.copyfile(uuid_file, mount_point+uuid_file)
+shutil.copyfile(uuid_file, mount_point_A+uuid_file)
 
-run_command('mkdir -p {0}/usr/lib/waggle && cd {0}/usr/lib/waggle && git clone https://github.com/waggle-sensor/waggle_image.git'.format(mount_point))
-
-### Create the image build script ###
-if is_extension_node:
-    local_build_script_in = 'extension_node_build.in'
-else:
-    local_build_script_in = 'nodecontroller_build.in'
-write_build_script('%s/root/build_image.sh' % (mount_point), local_build_script_in)
-
-print "-------------------------\n"
-with open('%s/root/build_image.sh' % (mount_point)) as local_build_script:
-  for line in local_build_script:
-    print(line)
-print "-------------------------\n"
-
-run_command('chmod +x %s/root/build_image.sh' % (mount_point))
+### Copy the image build script ###
+build_script = '%/root/configure_waggle.sh' % mount_point_A
+shutil.copyfile('%s/scripts/configure_waggle.sh' % waggle_image_directory, build_script)
+run_command('chmod +x %s' % build_script)
 
 configure_aot = False
 if os.path.exists('/root/id_rsa_waggle_aot_config') and run_command('ssh -T git@github.com', die=False) == 1:
@@ -466,11 +454,11 @@ if configure_aot:
     run_command('git clone git@github.com:waggle-sensor/private_config.git', die=False)
 
     # allow the node setup script to change the root password to the AoT password
-    shutil.copyfile('/root/id_rsa_waggle_aot_config', '%s/root/id_rsa_waggle_aot_config' % (mount_point))
-    shutil.copyfile('/root/private_config/encrypted_waggle_password', '%s/root/encrypted_waggle_password' % (mount_point))
+    shutil.copyfile('/root/id_rsa_waggle_aot_config', '%s/root/id_rsa_waggle_aot_config' % (mount_point_A))
+    shutil.copyfile('/root/private_config/encrypted_waggle_password', '%s/root/encrypted_waggle_password' % (mount_point_A))
 
     # allow the node the register in the field
-    shutil.copyfile('/root/private_config/id_rsa_waggle_aot_registration', '%s/root/id_rsa_waggle_aot_registration' % (mount_point))
+    shutil.copyfile('/root/private_config/id_rsa_waggle_aot_registration', '%s/root/id_rsa_waggle_aot_registration' % (mount_point_A))
   except Exception as e:
     print("Error in private AoT configuration: %s" % str(e))
     pass
@@ -488,7 +476,7 @@ print("Additional Pre-chroot Setup Duration: %ds" % (pre_chroot_time - loop_moun
 print "################### start of chroot ###################"
 
 if debug == 0:
-    run_command('chroot %s/ /bin/bash /root/build_image.sh' % (mount_point))
+    run_command('chroot %s/ /bin/bash /root/configure_waggle.sh' % (mount_point_A))
 
 
 print "################### end of chroot ###################"
@@ -503,43 +491,39 @@ print("Chroot Node Setup Duration: %ds" % (chroot_setup_time - pre_chroot_time))
 # After changeroot
 #
 
-if configure_aot:
+if configure_aot and not is_extension_node:
   # install a copy of wvdial.conf with the AoT secret APN
-  shutil.copyfile('/root/private_config/wvdial.conf', '%s/etc/wvdial.conf' % (mount_point))
+  shutil.copyfile('/root/private_config/wvdial.conf', '%s/etc/wvdial.conf' % (mount_point_A))
 
   # remove temporary password setup files from image
-  os.remove('%s/root/id_rsa_waggle_aot_config' % (mount_point))
-  os.remove('%s/root/encrypted_waggle_password' % (mount_point))
+  os.remove('%s/root/id_rsa_waggle_aot_config' % (mount_point_A))
+  os.remove('%s/root/encrypted_waggle_password' % (mount_point_A))
 
   # remove the private_config repository
   shutil.rmtree('/root/private_config')
 else:
   # copy the default, unconfigured wvdial.conf file
-  shutil.copyfile(waggle_image_directory + '/device_rules/wwan_modems/wvdial.conf', '%s/etc/wvdial.conf' % (mount_point))
+  shutil.copyfile(waggle_image_directory + '/device_rules/wwan_modems/wvdial.conf', '%s/etc/wvdial.conf' % (mount_point_A))
 
 
 try:
-    os.remove(new_image+'.report.txt')
+    os.remove(new_image_a+'.report.txt')
 except:
     pass
 
-print "copy: ", mount_point+'/'+report_file, new_image+'.report.txt'
+print "copy: ", mount_point_A+'/'+report_file, new_image_a+'.report.txt'
 
-if os.path.exists(mount_point+'/'+report_file):
-    shutil.copyfile(mount_point+'/'+report_file, new_image+'.report.txt')
+if os.path.exists(mount_point_A+'/'+report_file):
+    shutil.copyfile(mount_point_A+'/'+report_file, new_image_a+'.report.txt')
 else:
-    print "file not found:", mount_point+'/'+report_file
-
+    print "file not found:", mount_point_A+'/'+report_file
 
 
 
 if is_extension_node:
-    write_file(mount_point+'/etc/network/interfaces', guest_node_etc_network_interfaces_d)
+    write_file(mount_point_A+'/etc/network/interfaces', guest_node_etc_network_interfaces_d)
 else:
-    write_file(mount_point+'/etc/network/interfaces', nodecontroller_etc_network_interfaces_d)
-
-
-
+    write_file(mount_point_A+'/etc/network/interfaces', nodecontroller_etc_network_interfaces_d)
 
 
 
@@ -547,11 +531,11 @@ old_partition_size_kb=int(get_output('df -BK --output=size /dev/loop0p2 | tail -
 print "old_partition_size_kb: ", old_partition_size_kb
 
 
-unmount_mountpoint(mount_point)
+unmount_mountpoint(mount_point_A)
 time.sleep(3)
 destroy_loop_devices()
 time.sleep(3)
-create_loop_devices(new_image, 0,  None, start_block_data)
+create_loop_devices(new_image_a, 0,  None, start_block_data)
 time.sleep(3)
 print "filesystem check on /dev/loop0p2 after chroot"
 check_partition(0)
@@ -592,7 +576,7 @@ check_time = time.time()
 print("Partition Check Duration: %ds" % (check_time - expansion_time))
 ####################
 
-sector_size=int(get_output('fdisk -lu {0} | grep "Sector size" | grep -o ": [0-9]*" | grep -o "[0-9]*"'.format(new_image)))
+sector_size=int(get_output('fdisk -lu {0} | grep "Sector size" | grep -o ": [0-9]*" | grep -o "[0-9]*"'.format(new_image_a)))
 
 
 front_size_kb = sector_size * start_block_data/ 1024
@@ -612,7 +596,7 @@ if new_partition_size_kb < old_partition_size_kb:
     ### fdisk (shrink partition)
     # fdisk: (d)elete partition 2 ; (c)reate new partiton 2 ; specify start position and size of new partiton
 
-    run_command('echo -e "d\n2\nn\np\n2\n%d\n+%dK\nw\n" | fdisk %s' % (start_block_data, new_partition_size_kb, new_image))
+    run_command('echo -e "d\n2\nn\np\n2\n%d\n+%dK\nw\n" | fdisk %s' % (start_block_data, new_partition_size_kb, new_image_a))
 
 
 
@@ -623,7 +607,7 @@ if new_partition_size_kb < old_partition_size_kb:
     #set -e
 
     # does not show the new size
-    #fdisk -lu ${new_image}
+    #fdisk -lu ${new_image_a}
 
     # shows the new size (-b for bytes)
     #partx --show /dev/loop0p2 (fails)
@@ -672,7 +656,7 @@ print("Boot Partition Check Duration: %ds" % (check3_time - check2_time))
 
 
 # write image to file
-run_command('pv -per --width 80 --size %d -f %s | dd bs=1M iflag=fullblock count=%d | xz -1 --stdout - > %s.xz_part' % (combined_size_bytes, new_image, blocks_to_write, new_image))
+run_command('pv -per --width 80 --size %d -f %s | dd bs=1M iflag=fullblock count=%d | xz -1 --stdout - > %s.xz_part' % (combined_size_bytes, new_image_a, blocks_to_write, new_image_a))
 
 ###### TIMING ######
 image_write_time = time.time()
@@ -680,31 +664,104 @@ print("New Image Write Duration: %ds" % (image_write_time - check3_time))
 ####################
 
 # test if file was compressed correctly
-run_command('unxz -t %s.xz_part' % format(new_image))
+run_command('unxz -t %s.xz_part' % format(new_image_a))
 
 try:
-    os.remove(new_image_compressed)
+    os.remove(new_image_a_compressed)
 except:
     pass
 
-os.rename(new_image+'.xz_part',  new_image_compressed)
+os.rename(new_image_a+'.xz_part',  new_image_a_compressed)
+
+
+###### TIMING ######
+compression_check_time = time.time()
+print("New Image Compression Check Duration: %ds" % (compression_check_time - image_write_time))
+####################
+
+
+
+
+# create second dd with different UUIDs
+if create_b_image:
+    
+    
+    if not os.path.isfile(change_partition_uuid_script):
+        print change_partition_uuid_script, " not found"
+        sys.exit(1)
+    
+    try:
+        os.remove(new_image_b+'.part')
+    except:
+        pass
+    
+    #copy image a to b
+    run_command('pv -per --width 80 --size %d -f %s | dd bs=1M iflag=fullblock count=%d  > %s.part' % (combined_size_bytes, new_image_a, blocks_to_write, new_image_b))
+    
+    # delete old b if it exists
+    try:
+        os.remove(new_image_b)
+    except:
+        pass
+
+    try:
+        os.remove(new_image_b+'.xz')
+    except:
+        pass
+
+    
+    os.rename(new_image_b+'.part',  new_image_b)
+    
+    # create loop device
+    create_loop_devices(new_image_b, 1,  None, None)
+    
+    # change UUID
+    run_command(change_partition_uuid_script+ ' /dev/loop1')
+    
+    
+    # the recovery image on the eMMC needs space:
+    
+    new_image_a_compressed_size = os.path.getsize(new_image_a_compressed)
+    new_image_a_compressed_size_mb = new_image_a_compressed_size / 1048576
+    
+    # increase partition size again
+    run_command('dd if=/dev/zero bs=1MiB of={} conv=notrunc oflag=append count={}'.format(new_image_b, new_image_a_compressed_size_mb+50))
+    time.sleep(1)
+    
+    
+    # verify file system
+    run_command('e2fsck -f -y /dev/loop1p2')
+    
+    time.sleep(1)
+    
+    # make filesystem use new space
+    run_command('resize2fs /dev/loop1p2')
+    time.sleep(1)
+    
+    
+    # compress 
+    run_command('xz -1 '+new_image_b)
+    
+    
+    #run_command('pv -per --width 80 --size %d -f %s | dd bs=1M iflag=fullblock count=%d | xz -1 --stdout - > %s.xz_part' % (combined_size_bytes, new_image_a, blocks_to_write, new_image_b))
+    #os.rename(new_image_b+'.xz_part',  new_image_b+'.xz')    
 
 
 ###### TIMING ######
 bimage_time = time.time()
-print("\"B\" Image Creation Duration: %ds" % (bimage_time - image_write_time))
+print("\"B\" Image Creation Duration: %ds" % (bimage_time - compression_check_time))
 ####################
 
 #
 # Upload files to waggle download directory
 #
 if not configure_aot and os.path.isfile( data_directory+ '/waggle-id_rsa'):
-    remote_path = '/mcs/www.mcs.anl.gov/research/projects/waggle/downloads/waggle_images/base/'.format(image_type, odroid_model)
+    remote_path = '/mcs/www.mcs.anl.gov/research/projects/waggle/downloads/waggle_images/{0}/{1}/'.format(image_type, odroid_model)
     scp_target = 'waggle@terra.mcs.anl.gov:' + remote_path
-    run_command('md5sum $(basename {0}.xz) > {0}.xz.md5sum'.format(new_image) ) 
+    run_command('md5sum $(basename {0}.xz) > {0}.xz.md5sum'.format(new_image_a) ) 
     
     
-    cmd = 'scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.xz {1}.xz.md5sum {2}'.format(data_directory, new_image, scp_target)
+    cmd = 'scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.xz {1}.xz.md5sum {2}'.format(data_directory, new_image_a, scp_target)
     
     count = 0
     while 1:
@@ -732,14 +789,19 @@ if not configure_aot and os.path.isfile( data_directory+ '/waggle-id_rsa'):
   
     run_command('echo "{0}" > {1}/latest.txt'.format(new_image_base +".img.xz", data_directory))
     run_command('scp -o "StrictHostKeyChecking no" -i {0}/waggle-id_rsa {0}/latest.txt {1}/'.format(data_directory, scp_target))
+  
+    if os.path.isfile( new_image_b+'.xz'):
+        # upload second image with different UUID's
+        run_command( 'md5sum $(basename {0}.xz) > {0}.xz.md5sum'.format(new_image_b))
+        run_command('scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.xz {1}.xz.md5sum {2}'.format(data_directory, new_image_b, scp_target))
 
   
-    if os.path.isfile( new_image+'.report.txt'): 
-        run_command('scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.report.txt {2}'.format(data_directory, new_image,scp_target))
+    if os.path.isfile( new_image_a+'.report.txt'): 
+        run_command('scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.report.txt {2}'.format(data_directory, new_image_a,scp_target))
       
   
-    if os.path.isfile( new_image+'.build_log.txt'): 
-        run_command('scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.build_log.txt {2}'.format(data_directory, new_image,scp_target))
+    if os.path.isfile( new_image_a+'.build_log.txt'): 
+        run_command('scp -o "StrictHostKeyChecking no" -v -i {0}/waggle-id_rsa {1}.build_log.txt {2}'.format(data_directory, new_image_a,scp_target))
 
 ###### TIMING ######
 upload_time = time.time()
